pyproject.toml
[tool.poetry]
name = "cor-pass"
version = "0.1.0"
description = ""
authors = ["ZhovtenkoA <zhowtenko.oleksiy@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.9"
fastapi = "^0.111.0"
fastapi-limiter = "^0.1.6"
uvicorn = "^0.30.1"
sqlalchemy = "^2.0.31"
pydantic = "^2.8.2"
fastapi-mail = "^1.4.1"
pydantic-core = "^2.20.1"
httpx = "^0.27.0"
pycryptodome = "^3.20.0"
black = "^24.4.2"
alembic = "^1.13.2"
psycopg2 = "^2.9.9"
psycopg2-binary = "^2.9.9"
pydantic-settings = "^2.3.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
